+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This file is used to list changes made in each version of cookbook-openstack-ide
 * Allow existing users to have their passwords updated properly
 * Bump Chef gem to 11.16
 * Add test to verify each endpoint can be configured seperatly
+* Update endpoint when endpoint for one service type exists
 
 ## 10.0.0
 * Upgrading to Juno